+++ testing/src/main/java/demos/Demo3UserInterface.java
@@ -1,23 +1,30 @@
 package demos;
 
-import dev.gamekit.core.*;
+import dev.gamekit.core.Application;
+import dev.gamekit.core.IO;
+import dev.gamekit.core.Scene;
 import dev.gamekit.ui.Spacing;
 import dev.gamekit.ui.enums.Alignment;
 import dev.gamekit.ui.enums.CrossAxisAlignment;
 import dev.gamekit.ui.enums.MainAxisAlignment;
+import dev.gamekit.ui.events.MouseEvent;
 import dev.gamekit.ui.widgets.Button;
 import dev.gamekit.ui.widgets.Image;
 import dev.gamekit.ui.widgets.*;
+import dev.gamekit.utils.Config;
+import dev.gamekit.utils.Resolution;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
+import java.util.Objects;
 
 import static dev.gamekit.ui.widgets.AlignParam.horizontalAlignment;
 import static dev.gamekit.ui.widgets.AlignParam.verticalAlignment;
 import static dev.gamekit.ui.widgets.ButtonParam.mouseListener;
-import static dev.gamekit.ui.widgets.ButtonParam.*;
+import static dev.gamekit.ui.widgets.ButtonParam.spacing;
 import static dev.gamekit.ui.widgets.FlexParam.*;
 import static dev.gamekit.ui.widgets.ImageParam.image;
+import static dev.gamekit.ui.widgets.MultiChildParentParam.children;
 import static dev.gamekit.ui.widgets.PaddingParam.padding;
 import static dev.gamekit.ui.widgets.SingleChildParentParam.child;
 import static dev.gamekit.ui.widgets.SizedParam.height;
@@ -25,139 +32,152 @@ import static dev.gamekit.ui.widgets.SizedParam.width;
 import static dev.gamekit.ui.widgets.TextParam.*;
 
 public class Demo3UserInterface extends Scene {
-  int x = 0, y = 0;
-  String text = "Kwame";
-  BufferedImage squareImage = IO.getResourceImage("square-img.jpg");
-  BufferedImage wideImage = IO.getResourceImage("wide-img.jpg");
-  BufferedImage btnBgImage = IO.getResourceImage("btn-bg.png");
-  BufferedImage btnHoverImage = IO.getResourceImage("btn-hover.png");
+  BufferedImage backdrop = IO.getResourceImage("planetfall-artwork.jpg");
+  BufferedImage icon = IO.getResourceImage("planetfall-logo.png");
+  BufferedImage scrim = IO.getResourceImage("transparent-black.png");
 
   public Demo3UserInterface() {
     super("Main Scene");
   }
 
   public static void main(String[] args) {
-    Application game = new Application("Demo 3 - Declarative UI") { };
+    Application game = new Application(
+      new Config("Demo 3 - Declarative UI", Resolution.HD, false)
+    ) { };
     game.loadScene(new Demo3UserInterface());
     game.run();
   }
 
   @Override
-  public void onUpdate() {
-    super.onUpdate();
-
-    if (Input.isKeyJustPressed(Input.KEY_SPACE)) {
-      updateUI(() -> {
-        text = "Kwame";
-        x = y = 50;
-      });
-    } else if (Input.isKeyJustReleased(Input.KEY_SPACE)) {
-      updateUI(() -> {
-        text = "Opare";
-        x = y = 0;
-      });
-    }
-  }
-
-  @Override
-  public void onRender() {
-    super.onRender();
-    // Clear the screen with black
-    Renderer.setColor(Color.BLACK);
-    Renderer.clear();
-  }
-
-  @Override
   public Widget onCreateUI() {
-    return Align.create(
-      horizontalAlignment(Alignment.CENTER),
-      verticalAlignment(Alignment.CENTER),
-      child(
-        Sized.create(
-          width(600),
-          height(480),
+    return Stack.create(
+      children(
+        Image.create(
+          image(backdrop)
+        ),
+        Align.create(
+          horizontalAlignment(Alignment.START),
+          verticalAlignment(Alignment.START),
           child(
-            Column.create(
-              mainAxisAlignment(MainAxisAlignment.SPACE_BETWEEN),
-              crossAxisAlignment(CrossAxisAlignment.CENTER),
-              gapSize(10),
-              children(
-                Padding.create(
-                  padding(new Spacing(x)),
+            Padding.create(
+              padding(new Spacing(24, 0)),
+              child(
+                Sized.create(
+                  width(480),
+                  height(480),
                   child(
-                    Stack.create(
+                    Image.create(
+                      image(icon)
+                    )
+                  )
+                )
+              )
+            )
+          )
+        ),
+        Align.create(
+          horizontalAlignment(Alignment.START),
+          verticalAlignment(Alignment.CENTER),
+          child(
+            Padding.create(
+              padding(new Spacing(256, 0, 0, 96)),
+              child(
+                Column.create(
+                  mainAxisAlignment(MainAxisAlignment.START),
+                  crossAxisAlignment(CrossAxisAlignment.STRETCH),
+                  gapSize(24),
+                  children(
+                    MainMenuButton.create("Tutorial",
+                      e -> System.out.println("0: " + e.type)),
+                    MainMenuButton.create("New Planet",
+                      e -> System.out.println("1: " + e.type)),
+                    MainMenuButton.create("New Campaign", null),
+                    MainMenuButton.create("Load Game", null),
+                    MainMenuButton.create("Online Multiplayer", null),
+
+                    Column.create(
+                      mainAxisAlignment(MainAxisAlignment.START),
+                      crossAxisAlignment(CrossAxisAlignment.START),
+                      gapSize(12),
                       children(
-                        Sized.create(
-                          width(300),
-                          height(150),
-                          child(
-                            Image.create(
-                              image(wideImage)
-                            )
-                          )
-                        ),
-                        Sized.create(
-                          width(400),
-                          height(200),
-                          child(
-                            Image.create(
-                              image(squareImage)
-                            )
-                          )
-                        ),
-                        Sized.create(
-                          width(150),
-                          height(60),
-                          child(
-                            Button.create(
-                              defaultBackground(btnBgImage),
-                              hoverBackground(btnHoverImage),
-                              pressedBackground(btnBgImage),
-                              spacing(new Spacing(10, 8, 25, 8)),
-                              mouseListener((e) -> System.out.println(e.type)),
-                              child(
-                                Text.create(
-                                  text("Click Me")
-                                )
-                              )
-                            )
-                          )
-                        )
+                        SubMenuButton.create("Commander Customization"),
+                        SubMenuButton.create("Options"),
+                        SubMenuButton.create("Credits"),
+                        SubMenuButton.create("Exit Game")
                       )
+                    )
+                  )
+                )
+              )
+            )
+          )
+        ),
+        Align.create(
+          verticalAlignment(Alignment.START),
+          child(
+            Sized.create(
+              width(Resolution.NATIVE.width()),
+              height(128),
+              child(
+                Image.create(
+                  image(scrim)
+                )
+              )
+            )
+          )
+        ),
+        Align.create(
+          verticalAlignment(Alignment.END),
+          child(
+            Stack.create(
+              children(
+                Sized.create(
+                  width(Resolution.NATIVE.width()),
+                  height(128),
+                  child(
+                    Image.create(
+                      image(scrim)
                     )
                   )
                 ),
                 Sized.create(
-                  width(128),
-                  height(80),
+                  width(Resolution.NATIVE.width()),
+                  height(128),
                   child(
-                    Text.create(
-                      text("Hello World"),
-                      alignment(Alignment.CENTER),
-                      vAlignment(Alignment.END),
-                      shadowEnabled(true),
-                      shadowColor(Color.GRAY),
-                      shadowOffsetX(10),
-                      shadowOffsetY(10),
-                      fontStyle(Font.BOLD)
-                    )
-                  )
-                ),
-                Row.create(
-                  mainAxisAlignment(MainAxisAlignment.SPACE_BETWEEN),
-                  crossAxisAlignment(CrossAxisAlignment.STRETCH),
-                  gapSize(10),
-                  children(
-                    Text.create(
-                      text(text),
-                      fontSize(24)
-                    ),
-                    Padding.create(
-                      padding(new Spacing(x)),
-                      child(
-                        Text.create(
-                          text("Another Text"),
-                          color(Color.CYAN)
+                    Row.create(
+                      mainAxisAlignment(MainAxisAlignment.END),
+                      crossAxisAlignment(CrossAxisAlignment.CENTER),
+                      gapSize(24),
+                      children(
+                        Button.create(
+                          spacing(new Spacing(12)),
+                          child(
+                            Padding.create(
+                              padding(new Spacing(12, 12, 18, 12)),
+                              child(
+                                Text.create(
+                                  text("Create Account"),
+                                  fontSize(12),
+                                  fontStyle(Font.BOLD)
+                                )
+                              )
+                            )
+                          )
+                        ),
+                        Button.create(
+                          spacing(new Spacing(12)),
+                          child(
+                            Padding.create(
+                              padding(new Spacing(12, 12, 18, 12)),
+                              child(
+                                Text.create(
+                                  text("Login"),
+                                  fontSize(12),
+                                  fontStyle(Font.BOLD)
+                                )
+                              )
+                            )
+                          )
                         )
                       )
                     )
@@ -169,5 +189,85 @@ public class Demo3UserInterface extends Scene {
         )
       )
     );
+  }
+
+  static class MainMenuButton extends Compose {
+    protected final String text;
+
+    public MainMenuButton(String text, MouseEvent.Listener mouseListener) {
+      super(
+        Button.create(
+          mouseListener(mouseListener),
+          child(
+            Padding.create(
+              padding(new Spacing(12, 12, 16, 12)),
+              child(
+                Text.create(
+                  text(text),
+                  fontSize(20),
+                  fontStyle(Font.BOLD)
+                )
+              )
+            )
+          )
+        )
+      );
+
+      this.text = text;
+    }
+
+    public static MainMenuButton create(
+      String text,
+      MouseEvent.Listener mouseListener) {
+      return new MainMenuButton(text, mouseListener);
+    }
+
+    @Override
+    public boolean stateEquals(Widget widget) {
+      if (widget instanceof MainMenuButton mainMenuButton) {
+        return Objects.equals(text, mainMenuButton.text);
+      }
+
+      return false;
+    }
+  }
+
+  static class SubMenuButton extends Compose {
+    protected final String text;
+
+    public SubMenuButton(String text) {
+      super(
+        Button.create(
+          spacing(new Spacing(12)),
+          child(
+            Padding.create(
+              padding(new Spacing(6, 6, 12, 6)),
+              child(
+                Text.create(
+                  text(text),
+                  fontSize(12),
+                  fontStyle(Font.BOLD)
+                )
+              )
+            )
+          )
+        )
+      );
+
+      this.text = text;
+    }
+
+    public static SubMenuButton create(String text) {
+      return new SubMenuButton(text);
+    }
+
+    @Override
+    public boolean stateEquals(Widget widget) {
+      if (widget instanceof SubMenuButton subMenuButton) {
+        return Objects.equals(text, subMenuButton.text);
+      }
+
+      return false;
+    }
   }
 }